Use more conservative alt-key detection
Occasionally terminals can send `\x1b[` which previously was being
treated as `Alt-[`, but this could also be the start of a mouse or
other escape sequence. After this update, tcell will wait for 200
milliseconds more when it finds a conflict to give the terminal
time to send more characters. This does result in a slight delay
when pressing a conflict key (only `Alt-[` to my knowledge) and
this means that tcell polls every 200 milliseconds which isn't
ideal.

Ref zyedidia/micro#779