fixes #191 tcell hangs in go1.10
This is not so much a fix, as a workaround for an incredibly busted
tty driver in macOS.  I've tried nearly every other approach I can
think of, and this is the best of a bunch of really rotten options.

The draw back here is that reinitializing the screen may not work as well
as we'd like, and that there may be a leaked descriptor and goroutine
after you Fini() on macOS.  Not many other approaches seem viable, and to
be honest, very few applications are likely to ever notice that.

If some Apple developer sees this, please talk to me about the fixes
needed in the tty driver -- this has been a known bug (to me at least)
for many years now.