An attempt to speed up tcell screen rendering.
I've noticed that my tcell applications can be quite slow updating the screen
over an ssh connection, especially compared to frameworks like
urwid. Debugging with strace showed that each cell was written to the terminal
interleaved with calls to epoll_pwait() e.g. here rendering two bytes, chosen
randomly from my strace capture:

9874  22:56:10.508632 write(5, "f", 1)  = 1
9871  22:56:10.508654 <... epoll_pwait resumed> [{EPOLLOUT, {u32=754322992, u64=139801939807792}}, {EPOLLOUT, {u32=754323200, u64=139801939808000}}], 128, -1, NULL, 3) = 2
9871  22:56:10.508674 epoll_pwait(4,  <unfinished ...>
9874  22:56:10.508806 write(5, " ", 1 <unfinished ...>
9871  22:56:10.508820 <... epoll_pwait resumed> [{EPOLLOUT, {u32=754322992, u64=139801939807792}}, {EPOLLOUT, {u32=754323200, u64=139801939808000}}], 128, -1, NULL, 3) = 2
9874  22:56:10.508847 <... write resumed> ) = 1
9871  22:56:10.508855 epoll_pwait(4,  <unfinished ...>

To render its screen, tcell processes each dirty cell in sequence - for each
cell it generates the appropriate bytes for the terminal, and then calls
io.WriteString() with the terminal file handle as an io.Writer and the cell's
rendered bytes. Behind the scenes, this calls File.Write() which itself calls
epipecheck() resulting in the epoll_pwait() syscall (from src/os/file.go). The
call to epipecheck() exists so that the runtime can terminate a program in the
case of EPIPE i.e. writing to a closed pipe (for more see
https://github.com/golang/go/issues/11845). The typical result is a syscall to
write 1 byte to the terminal followed by epoll_pwait(), for each dirty cell in
the screen.

This change coalesces the terminal file handle writes into 1 call to
io.WriteString() per screen render. The function drawCell() is renamed
renderCell() and returns two strings in addition to the rendered cell
width. The first string may contain "delay" codes like $<100> and should be
rendered with TPuts(). The second string is the representation of the cell
contents. Instead of sending them to the terminal file handle, now both
strings are concatenated to a bytes.Buffer, the first post-processed by
turning "delay" codes into padding bytes. Once each dirty cell has been
processed in this way, the bytes.Buffer is converted to a string and that
string is written to the terminal.

Here is an example of a simple tcell application before this change:
https://drive.google.com/open?id=1k1CjK3vF3yOyubUhGCaPxyNsUbSrDmWs

And here is the same application after the change:
https://drive.google.com/open?id=1ERdYcjN23-WPPk-Mf7imcljFYR2ShiF8

I simulated a slower-than-normal VPN connection using

$ tc qdisc add dev tun0 root netem delay 200ms