A second attempt to speed up tcell screen rendering.
I've noticed that my tcell applications can be quite slow updating the screen
over an ssh connection, especially compared to frameworks like
urwid. Debugging with strace showed that each cell was written to the terminal
interleaved with calls to epoll_pwait() e.g. here rendering two bytes, chosen
randomly from my strace capture:

9874  22:56:10.508632 write(5, "f", 1)  = 1
9871  22:56:10.508654 <... epoll_pwait resumed> [{EPOLLOUT, {u32=754322992, u64=139801939807792}}, {EPOLLOUT, {u32=754323200, u64=139801939808000}}], 128, -1, NULL, 3) = 2
9871  22:56:10.508674 epoll_pwait(4,  <unfinished ...>
9874  22:56:10.508806 write(5, " ", 1 <unfinished ...>
9871  22:56:10.508820 <... epoll_pwait resumed> [{EPOLLOUT, {u32=754322992, u64=139801939807792}}, {EPOLLOUT, {u32=754323200, u64=139801939808000}}], 128, -1, NULL, 3) = 2
9874  22:56:10.508847 <... write resumed> ) = 1
9871  22:56:10.508855 epoll_pwait(4,  <unfinished ...>

To render its screen, tcell processes each dirty cell in sequence - for each
cell it generates the appropriate bytes for the terminal, and then calls
io.WriteString() with the terminal file handle as an io.Writer and the cell's
rendered bytes. Behind the scenes, this calls File.Write() which itself calls
epipecheck() resulting in the epoll_pwait() syscall (from src/os/file.go). The
call to epipecheck() exists so that the runtime can terminate a program in the
case of EPIPE i.e. writing to a closed pipe (for more see
https://github.com/golang/go/issues/11845). The typical result is a syscall to
write 1 byte to the terminal followed by epoll_pwait(), for each dirty cell in
the screen.

My prior unsatisfactory commit turned the operation of drawing a Screen cell
into the rendering of two strings - the first string possibly containing
"delay" codes like $<100> to be rendered with terminfo's TPuts() - and the
second string with the representation of the cell contents. This change was
not ideal because in order to write the concatenation of all cells to the
terminal in one shot, it required exporting a function from the terminfo package
to translate these delay codes into bytes for the terminal. That function
should not be part of terminfo's public interface.

This second, simpler approach adds a byte buffer and a buffering flag (bool)
to tcell's Screen. If buffering is true, calls to TPuts() will write to the
byte buffer; if false, to the os.File representing the terminal. Next, the
Screen's draw() function is modified to turn on buffering before starting to
render each cell; turn off buffering after rendering each cell; then finally,
to send the Screen's buffer to the terminal's os.File in one write.

These examples below were constructed with unsatisfactory change #1, but the
effect is the same with this new proposed change.

Before:
https://drive.google.com/open?id=1k1CjK3vF3yOyubUhGCaPxyNsUbSrDmWs

After:
https://drive.google.com/open?id=1ERdYcjN23-WPPk-Mf7imcljFYR2ShiF8

I simulated a slower-than-normal VPN connection using

$ tc qdisc add dev tun0 root netem delay 200ms