Deliver tcell events more efficiently to tcell applications.
This patch introduces a new tcell event type, EventBundle, which comprises a
slice of tcell.Events. Prior to this change, tcell will deliver events one at
a time (key, mouse, etc) to applications over an event channel of type
Event. The receiving application can call screen.PollEvent(), perhaps in a
goroutine - that function will return each event in sequence, blocking if none
are present. The application might then issue a type switch on the concrete
type of Event, processing *EventKey, *EventMouse, etc.

One problem the receiving application has to manage is when to issue a
redraw. For an application to feel responsive, it should update its UI as
quickly as possible after handling input. To do this, the application can
redraw after each call to PollEvent(). But this might result in unnecessarily
many redraws - for example, if a user has pasted a block of text into a tcell
application with the mouse, the application would redraw after each
character. A more complicated approach would be to redraw a short time
interval after the last call to PollEvent() (via a separate goroutine), while
cancelling and restarting a timer for that interval if another PollEvent()
returns in the meantime. This could work, but choosing the delay length
requires a trade-off between UI responsiveness and minimizing redraws.

Another way to handle this would be for tcell to send more than one event at
once, if possible. For example, if multiple events are extracted from the tty
input stream in one call to scanInput(), tcell could send them in a single
bundle via the event channel. This patch makes that possible with the new
EventBundle type.

Existing tcell applications have been built to handle the current set of tcell
types that implement tcell.Event. Because they would not understand and
process the EventBundle type, this feature must be opt-in. This is handled via
new Ext() APIs for creating a tcell screen - NewTerminfoScreenExt() and
NewScreenExt(). These both return a Screen that is identical to that returned
by NewTerminfoScreen() and NewScreen() respectively except that the screen
will send events over the event channel in bundles if possible.