If the tcell app is made smaller, allow it to be enlarged again.
Commit 1097a48ec0820f2e35d068395e2d79048acfa025 changed the implementation of
the tcell windows console screen to use a new console screen buffer via
CreateConsoleScreenBuffer. The screen buffer is initialized with a size that
is determined by the console window. If the user uses a mouse to make the
window smaller, the next call to cScreen.Show() calls resize() which sets the
console screen buffer to the smaller size. This has the side effect of
disabling the ability of the mouse to make the window larger again. Although
the window can be enlarged via right-click->Properties on the window title
bar, it's inconvenient.

This change removes the call to adjust the size of the console screen
buffer. Tcell's rendering functions are aware of the reduced window size, and
operate correctly in the reduced dimensions, but the windows buffer object
keeps its original size. This preserves the ability of the user to grab the
bottom right corner of the window with the mouse and resize the window back up
to its original size.