Add interceptor and pixel APIs for raw terminal support
This commit adds several interfaces to allow raw terminal support, such
as for SIXEL graphics.

DrawInterceptor was added for extensions to intercept each draw event
to synchronize its state with the screen's and redraw things it needs to
while occupying the screen lock to prevent data race with the user.

PixelSizer was added to expose the screen's size in pixels from
TIOCGWINSZ. This is needed to properly resize and position SIXEL images
on the terminal.

CellBufferViewer was added for low-level efficient damage tracking for
draw interceptors to synchronize the screen state.