Convert documentation to Markdown.
While asciidoc is probably technically more flexible, some sites
(such as pkg.go.dev) do not support it well.  For the basic docs
we have here, the extra richness from asciidoc is not needed, and
we would prefer not to sacrifice the broader support that MarkDown
enjoys.

While here the text was cleaned up, made consistent, with grammar
and spelling fixes.

The Tutorial was revised somewhat to provide more accurate guidance
for readers.