terminfo: continue processing if using an alias
Currently, the dynamic terminfo loader will return an empty struct if it
encounters that an alias was used. An alias does not mean the output of
`infocmp` was incorrect. Set the name to the primary name from `infocmp`
and continue processing as necessary.

For users who have a terminal not in the internal terminfo database who
also may be using an alias TERM (for whatever reason...), the result is
that no error is given to the application, and the terminfo struct
returns empty strings at runtime...notably the application can't enter
the alternate screen and the app is immediately unusable.