feat: Modern XTerm CSI-u key reporting (fixes #671)
This now discriminates between e.g. Tab and Ctrl-I.  But it does require
support from the terminal emulator.  Many do.

Modifiers have changed (we added ModHyper) and some keys
have been added as well.  More will be added later.
With this improvement we actually have the ability to obtain modifiers
for some keys that we otherwise would not.

Windows key reporting is somewhat harmonized as well, so that we get as
close to the same experience on all platforms as possible.