test: Add a test to ensure that all key mappings are unambiguous.
This test is designed to prove that we can potentially replace
the entire terminfo scheme for processing key codes with a set of
rich alternatives an ECMA-48 compliant input stream processor.

Also, while here, the KeyBackspace2 is converted to KeyBackspace
on input, so applications need only concern themselves with KeyBackspace.