chore(term): Remove hpterm.
This is a quasi ANSI terminal, but long ago fell out of use, as
more modern versions of HPUX support either dtterm (CDE) or some
other more main-stream desktop environment. That said HPUX itself
is reaching its final end-of-life in December 2025 (days away at
this writing), and there is no conceivable reason we should retain
support for the old hpterm.