chore(input): Complete refactoring of input processing.
This harmonizes the entire input process for VT style input
streams.  The intention here is to be able to use VT input on
Windows using win32-input-mode, but to share that code, and
eliminate as many of the terminal specific hacks as possible.

This relies on the fact that there are *almost* no conflicts in the
escape sequences used to report keys or input events.  The only
exception here is aixterm, which conflates CSI-P.