chore!: eliminate support for legacy Windows console
This converts us to use the new style Windows VT console (which is
found on Windows 10 (Creators Update and later, so anything after 2017)
and Windows 11.  Note that Windows 10 itself is EOL, with only ESU updates
possible. Presumably anyone with such systems will be running a newer edition.

This may have a deleterious impact on ConEmu, as the last time we tested it
it did not work well at all in VT emulation mode (this also greatly impaired
its usability using WSL).  Its somewhat likely that this is fixed in newer
editions, but we have not checked.