feat: Refactor Windows to use tscreen by default. (fixes #777)
This is a major redesign of the terminal subsystem for Windows to
utilize win32-input-mode and the common tscreen layer, so that we
can ultimatey (in followup) discard the console_win.go code entirely.
This also means that Windows can start benefitting from improvements
made to the common API.

One thing we should do is reduce the number of system calls used for
reading escape sequences (that will be in follow up).

This has been tested with Windows Terminal, Alacritty, WezTerm and
ConEmu.  We've encountered bugs in all of the 3rd party emulators,
and we cannot honestly recommend their use on Windows -- the Windows 11
Terminal application is however quite good and works very well with
this functionality. Alacritty performs acceptably, with workarounds,
WezTerm marginally so, and with ConEmu you should already be used to
to substantial limitations.

Follow up will be to improve performance and ultimately remove the
legacy console_win.go and support for NewConsoleScreen entirely.