Turn off CSI-u before leaving alternate screen mode
Terminals are required to remember CSI-u mode separately for the main screen and
the alternate screen, and maintain separate stacks of flags for each. This means
that it is important to disable CSI-u mode *before* we leave the alternate
screen; otherwise, it tries to pop the flags from the main screen stack, which
does nothing because it is empty, but leaves the flags on the alternate screen
stack forever, so they remain active for other, non-CSI-u aware applications.