feat: Add improved string accessors for cell content.
Having separate runes was awkward in many cases, and was done with
an incomplete understanding of unicode segmentation.  Its simpler for
everyone if we just treat these as strings, so we introduce new methods
for screen (Put, PutStyledStr, and PutStr) for writing either a grapheme
cluster, or (for convenience) writing a full string at a given location.

The other methods are now deprecated.  We intend to remove them in tcell v3.