fix(colors): Color was taking *far* too much space.
The size of color was 64-bits, in the hope that someday we would
be able to support an alpha channel.  That has never materialized,
and we're doubtful it is useful for body text.  Reducing colors to
32-bits cuts our memory usage in half for the screen buffer, and still
leaves four unused bits which could give up to 16 levels of alpha
if we ever need it.  (But most likely we would just add an 8-bit alpha
value for the background, as it seems unlikely that an alpha channel
for the foreground would ever be useful.)