chore!: Remove support for padding delay and pad characters.
This *might* break old VT100/VT102 physical terminals, but should
be an improvement for emulators. Actual Digital documentation and
ECMA-48 make no references to these delays, so they might not have been
needed anyway, but we're talkign about terminals made nearly 50 years
ago, and these are unlikely to exist (in any kind of functional state at
least) outside a museum.

Whereas the VT100 are functionality extinct, the VT100 terminal type
is still frequently used as a default in emulation, and can be found
in various places -- this change will make life better for them.