chore(term): Use VT100 character set for drawing fallback everywhere.
ANSI supported a different character set, in theory, but we think every
interesting terminal emulator at least supported the G0 set for line
drawing.

This also converts to using a cleaner solution for enabling the ACS,
using the same logic for all terminals. This might be slightly less efficient
for newer terminals, but if performance is a concern, use a UTF-8 capable
terminal.