fix: 3rd party Windows terminals double encoding win32-input-mode breaks some keys
The encoding for things like SS3 sequences used in terminals like WezTerm and Alacritty
breaks the processor, since the CSI state machine changes interferes with the SS3
state.  We introduce a nested state machine for this case.

Note that ConEmu still seems busted when running locally, although it is interesting
that using it from an SSH hosted session works well.  Nonetheless, we don't recommend
ConEmu for a variety of reasons.