feat: fallback to US-ASCII by default.
This allows most terminal applications that do not make use of Unicode
to Just Work.  (It does predispose for English.)  All supported encodings
are strict US-ASCII supersets, so this is entirely non-risky.

Applications which only need to work with newer systems could just assume
UTF-8 by setting the fallback to UTF-8 by default, but we didn't feel
confident of that yet.