refactor: defer registering resize handler (signal handler) until Notify call, remove extra locking
We need to move this logic later, because we plan to conditionalize this
so that we don't set up signal handling for terminals that can give us in-band
size updates.

This also eliminates the extra locks for the tty object, as terminal calls these functions with
the locks held  We only need them guarding the callback for the signal handler, as that may
be called asynchronously.