refactor!: Tty NotifyResize takes a channel instead of a function
This simplifies the shutdown quite a bit, as we can just stop
reading from the channel, and not worry about asynchronous go routines
running callback functions elsewhere.

This does represent a breaking change for anyone who has implemented
their own Tty, but we don't believe that case is typical.  Because
it is a breaking change, users should either fix their code, or
ensure that tcell is locked to an older version.

This falls quite on the heels of tcell 3.0, and we think adoption of
the new tcell v3 is still minimal, so we're accepting this breaking
change now, accepting the risk, but believing that nobody in actual
use will be negatively impacted.