feat: support inline resize and refactor private mode support
This adds a number new private mode values that we can query for,
and a cleaner internal API for setting and querying them.

The first use of this is to query for the ability to do inline
resize notifications, and if we have it, to skip use of signals.
This may allow us to work with resize events in circumstances where
the OS does not provide resize indications (for example if the
file descriptor in use not actually a tty!)  Only a few terminals
support this resize notifcation (ghostty, iTerm2, kitty, foot),
btu they are not niche.