test: create initial mock tty for testing
This is very primitive right now, but we'll add tests.  This MockTty
is, wait for it, a terminal emulator! (or simulator).

More tests will be added, and a full parser for incoming commands,
as well as responses from the terminal to queries made.  We will also
support injecting commands for the purposes of simulation.