fix: mouse support missing on macOS terminal
The macOS terminal does not support the DEC private mode queries,
so we cannot rely on it to discover SGR mice. While we'd like to
just tell users to complain to Apple or use another emulator, the
impact of this change on our downstreams is too large, so we need
to do this ugly workaround.

Note that there will still be some random "p" characters on screen,
as the Terminal does not fully eat the private mode sequences either.