chore!: Remove the views package
Nobody is really using this package, and we meant to do it for v3.
We think its soon enough after v3 released that its *probably* safe
to do this without cranking another major number.

The views content might come back in the form of another repository,
where hopefully it could evolve more freely, if there was interest in
doing so. (It's not immediately clear that sufficient interest exists.)