chore: move Tty to sub-subpackage (including WindowSize)
We also added compatibility aliases.  The purpose of this is to break
potential circular dependencies so that we can use symbols for TTYs
and the WindowSize in the vt emulator and mock packages.

We want to then be able to add tests in the main tcell package, which
can use these -- this move breaks the circular dependency.