chore: remove ineffectual assignments
The thinking was once to include the rune (string) for the control
key presses, but as we have not done that in the past, and as we have
test cases that verify that we don't, we will leave it alone for now.

In the future we will probably introduce additional reporting modes, and
we'll offer an opt-in to change this reporting in the future.