feat(emulator): Support emulator resizing
This adds automatic handling of inline notifications as well, and
adds both tscreen and emulator specific test cases.

In order to make sure that this code is bullet proof for asynchronous
event injection, we also made writing to the read queue atomic for the
full length of the written strings.