docs: improve the badges
We use a short-form ukraine badge (which is enough for now), remove
the license badge (the license is already obvious in GitHub's UI),
and eliminate the dynamic goreportcard badge, providing it with a static
one. (This sort of cheats, but we have been A+ ~forever, so it's not
lying, as long as we keep to the high standards.  If goreportcard ever
lets us customize the badge styling then we could go back to a dynamic version.)