refactor(emulator): Refactoring of the emulator in anticipation of grapheme cluster support
This has the initial effort for grapheme clusters, and also for performing operations like
scrolling, insertion, etc.  This requires that the emulator keep its own copy of the cell
data, so it can perform modifications without requiring the backend to report back the content.
(We ultimately wnat to be able to support fully stateless backends.)

To make things faster, we introduce an as yet unused Blitter interface, that lets us move
content on screen from one location to another.  (This operation should be familiar to anyone
who's worked with graphics systems.)

The Positioner interfaces are removed as well.  On consider we find that they are unlikely
to provide much use, and for backends (such as layering on top of a real terminal) that
have a stateful cursor, they can implement this trivially by comparing the position sent
with their own notion of the cursor.

Finally the attribute handling is moved to be more representative of a cursor, like we did
for colors, where we set the attributes once that will be used for subsequent operations.  This
makes for a simpler API all around.